9th nov-paet2

+++ Nov/ex_09112024)_Literals_variables/Lab022_User_Input_Sum_Of_Two_numbers.py
@@ -7,12 +7,19 @@
 #step 1
 #I/P -> num1,num2 ->int
 #O/P ->sum->int, sub->int, div->float
-num1 = int(input("entr the num 1"))
-num2 = int(input("entr the num 2"))
+
+#num1 = int(input("entr the num 1"))
+#num2 = int(input("entr the num 2"))
+
+num1 = float(input("entr the num 1"))
+num2 = float(input("entr the num 2"))
+
 # num1=int(num1)  #str->int
 # num2=int(num2)  #str->int
+
 print((type(num1)))
 print((type(num2)))
+
 #step 2/ Rough lOgic
 #+,-,/,*,
 #step 3
@@ -21,11 +28,13 @@ sum= num1 + num2
 sub= num1 - num2
 mul= num1 * num2
 div= num1 / num2
+
 print("sum is :", sum)
 print("sub is :", sub)
 print("mul is :", mul)
 print("div is :", div)
-# num1->155
+
+# num1->155 or 155.6
 #num2->5
